Split internal function into two
This splits out the portion of a function that calculates the index into
our tables from a locale category.  This immediately allows the calling
sequence to be simplified, but future commits will have use for just
the functionality of the new function.

One debugging statement is also removed from the new split out function,
which was not really necessary, and made it harder to do the next
commit.