Create record_state_of_cache() in ext/Pod-Html/t/lib/Testing.pm
This function permits examination of the state of the Pod-Html cache
file between calls to xconvert().  This is being added to diagnose the
problems reported in https://github.com/Perl/perl5/issues/12271.

2. Allow an 'expect_fail' option to be passed to xconvert().

This will be passed internally to _process_diff().  Within the latter,
if expect_fail is Perl-false, the unit test will FAIL and the diff
between the expectation and the result will be dumped as has always been
the case.  If, however, we expect the test to fail but want to indicate
a reason for that failure, we should (a) adjust the 'description' and
(b) set expect_fail to Perl-true.  When the calculated HTML fails to
match the expectation, the unit test will PASS but the description will
inform the user that we have an anomaly that has been diagnosed.

Adapt 3 t/*htmldir3.t test files to this interface modification.