Add two defective test files.
The purpose of this branch is to explore the reasons behind the test
failures observed in https://github.com/Perl/perl5/issues/12271.  In
this commit we simply add the two dummy test files which illustrate the
problem. xhthmldir3.t comments out the one good run; yhtmldir3.t
reverses the order of the two runs such that the bad one precedes the
good one.

We clearly know that one run is defective and will have to be deleted.
What we don't yet know is why it gives a spurious pass if it is preceded
by the good run.