Allow for expectation that xconvert() will fail.
Allow for an 'expect_fail' element to be passed to xconvert().
This will be passed internally to process_diff().  Within the latter, if
expect_fail is Perl-false, the unit test will FAIL and the diff between
the expectation and the result will be dumped as has always been the
case.

If, however, we expect the test to fail but want to indicate a reason
for that failure, we should (a) adjust the 'description' and (b) set
expect_fail to Perl-true.  When the calculated HTML fails to match the
expectation, the unit test will PASS but the description will inform the
user that we have an anomaly that has been diagnosed.

Adapt 3 t/*htmldir3.t test files to this interface modification.

After this commit I'll fork a new branch in which I start to clean up
the code and polish it up for a pull request.

Adapt other test files