Debugging code
This branch should end here.  I have been able to determine that we
should be able to safely eliminate 'htmlroot' and 'podroot' as KVPs with
_prepare_args_table(), as they do not affect what Pod::Html does
internally in any significant way.

On the other hand, 'podpath', if commented out, causes a failure in 1
file: t/htmlview.t.  I have not yet been able to figure this out.