Unlink the tempfile at end of each cycle
Once I dumped the atime (index 8) of the stat and lstat calls on the
write, I noticed that they were not advancing from one cycle to the
next.  I wondered whether this could be due to the fact that no 'unlink'
is called until after the 5 cycles have all completed.

If I 'unlink' the tempfile at the end of each cycle, the atime values
advance as expected.

All existing tests pass for me in either case.