threads and threads-shared: do not run
These two distros can only be run meaningfully against a threaded build.  Add
a column to the distro metadata files indicating 'needs_threads', then skip
this distro on all versions of perl if that column has true value.

At a later stage in this project we'll have to consider switching over to
Tux's threaded builds.  But there we'll face the problem that he doesn't use
symlinks to point to the last maintenance release of a given perl version as
he does with non-threaded builds.  That will require changes to the older
versions config file and to many points in the program; we'll have to use
3-part version numbers throughout.