Get some dist/threads/t/*.t tests running
Prior to this commit 'dist/threads' tests were failing for two distinct
reasons:

1. We're not yet setup to test with Tux's threaded perls.

2. Most of these tests are set up to test with t/test.pl rather than
Test::More.  t/test.pl therefore has to be locatable so it can be require-d in
a BEGIN block.  But we're testing in a tempdir, not in our git checkout.  So
we have to copy t/test.pl to an appropriate place in that tempdir.

This commit makes t/test.pl -- as it currently exists in blead -- available to
our tests in their tempdir.  However, the current syntax of t/test.pl
apparently can only be run in 5.14 and later.  So we have to supply a
minimum_perl_version for threads and threads-shared in
dist-backcompat-distro-metadata.txt.  There happens to be one test in
dist/threads that does *not* require a threaded build, so that test is now
runnable in 5.14 and later.

For the time being, we'll comment out the part of
test_one_distro_against_older_perls() that skips distros with a true value for
'needs_threads' in the metadata file.  Once we switch to using threaded
builds, we'll un-comment this part of the code.