Convert to using Tux's threaded perls
These will be found in /media/Tux/perls-t/bin instead of /media/Tux/perls/bin.
In this directory Tux does not use any symlinks pointing to the last
maintenance release in a given cycle.  So in the config file we have to spell
these like, say, 'perl5.18.4' rather than perl5.18. This also means we can
remove the code which checks for symlinks and targets, which simplifies the code.

In two places, correct failures to encapsulate code within a subroutine.

Curiously ...

Using threaded builds all versions of 'threads' still fail, but
'threads-shared' passes on 5.34.0.  Search-Dist now fails on 5.6.2.