OpenBSD: Force hints for dlopen and dlerror
DO NOT MERGE until 'echo' debugging statements in Configure have been
removed.

On OpenBSD, Configure was failing to detect dlopen(), which in turn
meant that it was failing to automatically find
ext/DynaLoader/dl_dlopen.xs when compiling with g++ and was therefore
pausing for manual user input (which, e.g., made smoke-testing
difficult).

We have apparently previously encountered this problem, as there is code
in hints/linux.sh to provide a work-around.  If we import this code into
hints/openbsd.sh, Configure runs to completion.

For: https://github.com/Perl/perl5/issues/21958