switch removal: ~~ no longer an op, so it still cannot chain
The test here does a simple eval EXPR to test whether $a op $b op $c
errors when trying to chain, and since ~~ is no longer an op, of
course the eval errors.

Unfortunately both failing to chain and the no longer valid op produce
the same error, so we can't easily make the test more robust by
validating the error message.