Standardize leading whitespace on 4
We'll need to do some refactoring and re-arranging of code in
make_ext.pl to address GH #24014.  That will be easier to do if we tidy
up the code inside sub build_extension():  300+ lines with a mixture of
hard tabs and whitespace for leading indents.  However, for the time
being we won't do additional tidying (e.g., uncuddling else's) because
that would change line numbers.